[bsp][stm32f429-disco] Change touch exec from hard timer to thread
