+++ bsp/stm32f429-disco/applications/application.c
@@ -32,8 +32,40 @@
 #include <gdb_stub.h>
 #endif
 
+#include "drv_touch.h"
+
+#ifdef PKG_USING_LITTLEVGL2RTT
+#include "littlevgl2rtt.h"
+#endif 
+
+static void rt_touch_thread_entry(void *parameter)
+{
+    int16_t x;
+    int16_t y;
+    struct touch_state ts;
+    while(1)
+    {
+        touch_get_state(&ts);
+
+#ifdef PKG_USING_LITTLEVGL2RTT    
+        if(ts.pressed) 
+        {
+            x = (3706 - ts.x) / 14;
+            y = (-461 + ts.y) / 10.5;
+        
+            littlevgl2rtt_send_input_event(x, y, LITTLEVGL2RTT_INPUT_DOWN);
+        }
+        else
+            littlevgl2rtt_send_input_event(-1, -1, LITTLEVGL2RTT_INPUT_UP);
+#endif
+        rt_thread_mdelay(100);
+    }
+}
+
 void rt_init_thread_entry(void* parameter)
 {
+    rt_thread_t tid;
+    
     /* GDB STUB */
 #ifdef RT_USING_GDB
     gdb_set_device("uart6");
@@ -58,6 +90,16 @@ void rt_init_thread_entry(void* parameter)
     
     rt_components_init();
 
+    rt_device_t tscreen = rt_device_find("touch");
+    rt_device_open(tscreen, RT_DEVICE_FLAG_RDWR);
+
+    tid = rt_thread_create("touch",
+                           rt_touch_thread_entry, RT_NULL,
+                           1024, 4, 20);
+
+    if (tid != RT_NULL)
+        rt_thread_startup(tid);
+                                      
 }
 
 int rt_application_init()
